In core: Make ShapeBasedConstraint child of Constraint.
